changes from cr
