+++ react-components/src/data-providers/types.ts
@@ -7,54 +7,11 @@ import {
   type ExternalId,
   type InternalId,
   type Metadata,
-  type Relationship,
-  type Node3D
+  type Relationship
 } from '@cognite/sdk/';
-import { type NodeItem, type DmsUniqueIdentifier, type Source, type EdgeItem } from './FdmSDK';
-import { type FdmCadConnection } from '../components/CacheProvider/types';
+import { type DmsUniqueIdentifier, type Source } from './FdmSDK';
 
 export type FdmInstanceWithView = DmsUniqueIdentifier & { view: Source };
-
-export type FdmInstanceWithProperties = NodeItem<unknown> | EdgeItem<unknown>;
-
-export type FdmInstanceWithPropertiesAndTyping = {
-  items: FdmInstanceWithProperties[];
-  typing: FdmTyping;
-};
-
-export type FdmTyping = Record<
-  string,
-  Record<
-    string,
-    Record<
-      string,
-      {
-        nullable?: boolean;
-        autoIncrement?: boolean;
-        defaultValue?: any;
-        description?: string;
-        name?: string;
-        immutable?: boolean;
-        type: { collation?: string; list?: boolean; type: string };
-      }
-    >
-  >
->;
-
-export type FdmInstanceNodeWithConnectionAndProperties = {
-  instanceType: 'node';
-  version: number;
-  space: string;
-  externalId: string;
-  createdTime: number;
-  lastUpdatedTime: number;
-  deletedTime: number;
-  items: FdmInstanceWithProperties[];
-  connection?: FdmCadConnection | undefined;
-  cadNode?: Node3D | undefined;
-  view?: Source | undefined;
-  typing: FdmTyping;
-};
 
 export type AssetInstanceReference = { assetId: number };
 export type InstanceReference = AssetInstanceReference | DmsUniqueIdentifier;
